refactor: move the test-seeding channel out of the example MainActivity
The example app is user-facing documentation, so keep MainActivity
trivially scannable: the sync-adapter seeding handler now lives in
TestSeedChannel.register(flutterEngine, contentResolver), called in
one line from configureFlutterEngine. No behavior change.
